correções no backend validação e autenticação. melhorias no frontend - autenticação, páginas de admin, de produtos, carrinho e mais

+++ api-js/src/modules/user/user.controller.js
@@ -1,24 +1,33 @@
 const service = require("./user.service");
 const dataValidator = require("../../common/validators/user.validator");
+const CustomError = require("../../common/utils/CustomError");
 
 module.exports = {
     async getUsers(req, res) {
+        if (!req.ability.can("read", "User")) throw new CustomError("Acesso negado!", 403);
+
         const users = await service.getUsers();
         return res.json(users);
     },
 
     async getUserById(req, res) {
+        if (!req.ability.can("read", "User")) throw new CustomError("Acesso negado!", 403);
+
         const user = await service.getUserById(req.params.id);
         return res.json(user);
     },
 
     async updateUser(req, res) {
+        if (!req.ability.can("manage", "User")) throw new CustomError("Acesso negado!", 403);
+
         const validateData = dataValidator.update(req.body);
         const updatedUser = await service.updateUser(req.params.id, validateData);
         return res.json(updatedUser);
     },
 
     async deleteUser(req, res) {
+        if (!req.ability.can("manage", "User")) throw new CustomError("Acesso negado!", 403);
+
         const user = await service.deleteUser(req.params.id);
         return res.json({ message: "Usuário deletado com sucesso", user });
     },
